requirements.txt
fastapi[standard]
cassandra-driver
PyJWT
passlib[bcrypt]
python-jose[cryptography] 
python-multipart